back/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
psycopg2-binary==2.9.10